% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laureates.R
\name{laureates}
\alias{laureates}
\title{Nobel laureate data}
\usage{
laureates(
  offset = NULL,
  limit = 25,
  sort = "asc",
  ID = NULL,
  name = NULL,
  gender = NULL,
  motivation = NULL,
  affiliation = NULL,
  residence = NULL,
  birthDate = NULL,
  birthDateTo = NULL,
  deathDate = NULL,
  deathDateTo = NULL,
  foundedDate = NULL,
  birthCity = NULL,
  birthCountry = NULL,
  birthContinent = NULL,
  deathCity = NULL,
  deathCountry = NULL,
  deathContinent = NULL,
  foundedCity = NULL,
  foundedCountry = NULL,
  foundedContinent = NULL,
  HeadquartersCity = NULL,
  HeadquartersCountry = NULL,
  HeadquartersContinent = NULL,
  nobelPrizeYear = NULL,
  yearTo = NULL,
  nobelPrizeCategory = NULL,
  csvLang = "en"
)
}
\arguments{
\item{offset}{The number of items to skip before starting to collect the result set. Minimum 1.}

\item{limit}{The numbers of items to return. Minimum 1. Default 25.}

\item{sort}{Sort order, result is sorted alphabetically by known name. Must be one of \code{c("asc", "des")}.}

\item{ID}{Numeric ID of the Laureate (unique key for each Nobel Laureate).}

\item{name}{Laureate's name (person or organization).}

\item{gender}{Laureate's gender if person. Must be one of \code{c("female", "male", "other")}.}

\item{motivation}{Text in Laureate's motivation}

\item{affiliation}{Affiliation(s) for the Laureate at the time of the award}

\item{residence}{Laureate's place of residence at the time is awarded}

\item{birthDate}{Birth date of Laureate if Person. Search by year in the form of YYYY, by month -MM-, by day -DD or YYYY-MM-DD}

\item{birthDateTo}{Return Laureates born within a range of years between birthDate and birthDateTo. BirthDate field is required}

\item{deathDate}{Death date of Laureate if Person. Search by year in the form of YYYY, by month -MM-, by day -DD or YYYY-MM-DD}

\item{deathDateTo}{Return Laureates deceased within a range of years between deathDate and deathhDateTo. DeathDate field is required}

\item{foundedDate}{Founding date of Laureate if Organization. Search by year in the form of YYYY, by month -MM-, by day -DD or YYYY-MM-DD}

\item{birthCity}{Laureate's city of birth if person}

\item{birthCountry}{Laureate's country of birth if person}

\item{birthContinent}{Laureate's continent of birth if person. Continent as in geonames.org standard. Must be one of \code{c("Africa", "Asia", "Europe", "North America", "Oceania", "South America", "Antarctica")}.}

\item{deathCity}{Laureate's city of death if person}

\item{deathCountry}{Laureate's country of death if person}

\item{deathContinent}{Laureate's continent of death if person. Continent as in geonames.org standard. Must be one of \code{c("Africa", "Asia", "Europe", "North America", "Oceania", "South America", "Antarctica")}.}

\item{foundedCity}{City where organization was founded.}

\item{foundedCountry}{Country where organization was founded.}

\item{foundedContinent}{Continent where organization was founded. Continent as in geonames.org standard. Must be one of \code{c("Africa", "Asia", "Europe", "North America", "Oceania", "South America", "Antarctica")}.}

\item{HeadquartersCity}{City where organization's hearquarters are.}

\item{HeadquartersCountry}{Country where organization's hearquarters are.}

\item{HeadquartersContinent}{Continent where organization's hearquarters are. Continent as in geonames.org standard. Must be one of \code{c("Africa", "Asia", "Europe", "North America", "Oceania", "South America", "Antarctica")}.}

\item{nobelPrizeYear}{Year the Nobel Prize was awarded, in the form of YYYY.}

\item{yearTo}{Used in combination with \code{nobelPrizeYear} to specify a range of years to return results from.\code{NobelPrizeYear} is required.}

\item{nobelPrizeCategory}{Nobel Prize Category. Must be one of \code{c("che", "eco", "lit", "pea", "phy", "med")}.}

\item{csvLang}{Language of output for csv format. Must be one of \code{c("en", "se", "no")}. Default = \code{"en"}.}
}
\value{
A data frame with one row for each result.
}
\description{
Nobel laureate data
}
