% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nobelPrizes.R
\name{nobelPrizes}
\alias{nobelPrizes}
\title{Nobel prize data
Obtain information about all Nobel Prizes or search for a specific
set of Nobel Prizes. Note that not all information about the Laureates
are provided in the output, as a request of making this endpoints response
lighter. Call the laureates endpoint for full information.}
\usage{
nobelPrizes(
  offset = NULL,
  limit = 25,
  sort = "asc",
  nobelPrizeYear = NULL,
  yearTo = NULL,
  nobelPrizeCategory = NULL,
  csvLang = "en"
)
}
\arguments{
\item{offset}{The number of items to skip before starting to collect the result set. Minimum 1.}

\item{limit}{The numbers of items to return. Minimum 1. Default 25.}

\item{sort}{Sort order, result is sorted by year. Must be one of \code{c("asc", "des")}.}

\item{nobelPrizeYear}{Year the Nobel Prize was awarded, in the form of YYYY.}

\item{yearTo}{Used in combination with \code{nobelPrizeYear} to specify a range of years to return results from.\code{NobelPrizeYear} is required.}

\item{nobelPrizeCategory}{Nobel Prize Category. Must be one of \code{c("che", "eco", "lit", "pea", "phy", "med")}.}

\item{csvLang}{Language of output for CSV format. Must be one of \code{c("en", "se", "no")}. Default = \code{"en"}.}
}
\value{
A data frame with one row for each result.
}
\description{
Nobel prize data
Obtain information about all Nobel Prizes or search for a specific
set of Nobel Prizes. Note that not all information about the Laureates
are provided in the output, as a request of making this endpoints response
lighter. Call the laureates endpoint for full information.
}
